third_party/blink/web_tests/external/wpt/css/selectors/parsing/parse-part-expected.txt

This is a testharness.js-based test.
[FAIL] "::part(foo):is(ul)" should be a valid selector
  assert_equals: serialization should be canonical expected "::part(foo):is(ul)" but got "::part(foo):is()"
[FAIL] "::part(foo):is(nav ul)" should be a valid selector
  assert_equals: serialization should be canonical expected "::part(foo):is(nav ul)" but got "::part(foo):is()"
[FAIL] "::part(foo):where(ul)" should be a valid selector
  assert_equals: serialization should be canonical expected "::part(foo):where(ul)" but got "::part(foo):where()"
[FAIL] "::part(foo):where(nav ul)" should be a valid selector
  assert_equals: serialization should be canonical expected "::part(foo):where(nav ul)" but got "::part(foo):where()"
Harness: the test ran to completion.
